LayoutTests/http/tests/inspector/network/resource-sizes-memory-cache-expected.txt

Test for Resource size values (transfer size, decoded size, header size, request and response) for resources served from the memory cache.


== Running test suite: Resource.Size.MemoryCache
-- Running test case: Resource.Size.MemoryCache.Basic
PASS: statusCode should be 304.
PASS: compressed should be false.
PASS: responseSource should be Symbol(memory-cache).
size: 27
requestBodyTransferSize: 0
responseBodyTransferSize: 0
estimatedNetworkEncodedSize: 0
networkEncodedSize: 0
PASS: estimatedTotalTransferSize should be >= (encoded body size + headers).
PASS: networkTotalTransferSize should be >= (encoded body size + headers).
PASS: requestHeadersTransferSize should be non-empty.
PASS: responseHeadersTransferSize should be non-empty.
